Initial Linked List Structure & Control Validator
Before initializing the snake movements, I created the Linked List structure to trace the whole snake body troughout the game board.
Although we have this game board implemented with the linked list, it might be useful for printing purposes only. This because we will do the game process fully by navigating trough the linked list that represents the snake.

I also created and tested the control validator. For this, i created a debug stringify function.

* Control validator: Checks the snake direction and validates if the next user input is valid or not.

Example: Valid if the snake is going up, and the user presses left or right. Not valid if the snake is going up and the user presses down, because its the oposite direction.