Output is Unicode UTF-8
Having few Windows Codepage characters to draw the game would not fullfill the requirements.

In order to draw the snake with full blocks, and the border with a proper outlining, I've changed the characters to UTF-8. This provides a wider selection of characters.

In order to do such improvement, I've changed the _setmode() and all the outputs are using wide strings.